+++ server/internal/handler/task/sorttaskhandler.go
@@ -1,0 +1,29 @@
+package task
+
+import (
+	"net/http"
+
+	"github.com/zeromicro/go-zero/rest/httpx"
+	"server/internal/pkg/xerr"
+	"server/internal/logic/task"
+	"server/internal/svc"
+	"server/internal/types"
+)
+
+func SortTaskHandler(svcCtx *svc.ServiceContext) http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		var req types.SortTaskReq
+		if err := httpx.Parse(r, &req); err != nil {
+			httpx.ErrorCtx(r.Context(), w, err)
+			return
+		}
+
+		l := task.NewSortTaskLogic(r.Context(), svcCtx)
+		resp, err := l.SortTask(&req)
+		if err != nil {
+			httpx.ErrorCtx(r.Context(), w, err)
+		} else {
+			httpx.OkJsonCtx(r.Context(), w, xerr.SuccessResponse(resp))
+		}
+	}
+}